+++ lord-save-me/src/App.js
@@ -1,20 +1,22 @@
-import './App.css';
-import Home from './pages/Home/Home';
+import { useState } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import NavBar from './components/NavBar/NavBar';
+
+//HOME PAGE
+import Home from './pages/Home/Home';
+import Policy from './pages/Home/Policy';
 import Forms from './pages/Register/Forms';
-import ClientLogin from './pages/Login/ClientLogin';
-import AdminLogin from './pages/Login/AdminLogin';
-import Policy from './pages/Policy';
+import Login from './pages/Login/Login';
 import SetPassword from './pages/Password/SetPassword';
 import ForgotPassword from './pages/Password/ForgotPassword';
 import OTP from './pages/Password/OTP';
-import Test from './pages/Register/Test';
-import Practice from './pages/Register/Practice';
-import BankDetails from "./pages/UserPanel/BankDetails";
-import { useState } from 'react';
 
+//COMPONENTS
+import NavBar from './components/NavBar/NavBar';
 import SideBar from './components/Sidebar/SideBar';
+import AdminSideBar from './components/Sidebar/AdminSideBar';
+import PostLoginNavBar from './components/NavBar/PostLoginNavBar';
+
+//USER PANEL
 import Dashboard from "./pages/UserPanel/Dashboard";
 import KYC from "./pages/UserPanel/KYC";
 import CheckStatus from "./pages/UserPanel/CheckStatus";
@@ -22,19 +24,26 @@ import AddInquiry from './pages/UserPanel/AddInquiry';
 import Analytics from "./pages/UserPanel/Analytics";
 import PersonalInfo from './pages/UserPanel/PersonalInfo';
 import LoanInfo from './pages/UserPanel/LoanInfo';
+import BankDetails from "./pages/UserPanel/BankDetails";
 
+//ADMIN PANEL
 import AdminDashboard from "./pages/AdminPanel/AdminDashboard.jsx";
 import ClientData from "./pages/AdminPanel/ClientData";
 import ClientDetails from "./pages/AdminPanel/ClientDetails";
 import Statistics from "./pages/AdminPanel/Statistics";
 
-import AdminSideBar from './components/Sidebar/AdminSideBar';
-import PostLoginNavBar from './components/NavBar/PostLoginNavBar';
+//TESTING
+import Test from './pages/Register/Test';
+import Practice from './pages/Register/Practice';
+import SideBar2 from './components/SideBar2/SideBar2';
+
+
 function App() {
   const [loginToken, setLoginToken] = useState({})
   return (
     <>
       <Routes>
+        {/* HOME */}
         <Route exact path='/' element={<Home />} />
         <Route path='loan/business' element={<Forms loan_type={"Business"} country={"India"} />} />
         <Route path='loan/home' element={<Forms loan_type={"Home"} country={"India"} />} />
@@ -42,11 +51,18 @@ function App() {
         <Route path='loan/personal' element={<Forms loan_type={"Personal"} country={"India"} />} />
         <Route path='/otp' element={<OTP />} />
         <Route path='/set-password' element={<SetPassword />} />
-        <Route path='/login/client' element={<ClientLogin />} />
+        <Route path='/login/client' element={<Login />} />
+        <Route path='/login/admin' element={<Login />} />
         <Route path='/forgot-password/:id' element={<ForgotPassword />} />
-        <Route path='/test' element={<Test />} />
-        <Route path='/practice' element={<Practice />} />
+        <Route path='/policy' element={<Policy />} />
+
+        {/* COMPONENTS */}
+        <Route path='/navbar' element={<NavBar />} />
         <Route path='/sidebar' element={<SideBar />} />
+        <Route path='/new-navbar' element={<PostLoginNavBar />} />
+        <Route path='/admin-sidebar' element={<AdminSideBar />} />
+
+        {/* USER PANEL */}
         <Route path="/sidebar/dashboard" element={<Dashboard />} />
         <Route path="/sidebar/kyc-docs" element={<KYC />} />
         <Route path="/sidebar/analytics" element={<Analytics />} />
@@ -55,15 +71,17 @@ function App() {
         <Route path="/sidebar/my-info/personal-info" element={<PersonalInfo />} />
         <Route path="/sidebar/my-info/loan-info" element={<LoanInfo />} />
         <Route path="/sidebar/bank-details" element={<BankDetails />} />
-        <Route path='/new-navbar' element={<PostLoginNavBar />} />
-        <Route path='/sidebar' element={<SideBar />} />
-        <Route path='/admin-sidebar' element={<AdminSideBar />} />
+
         {/* ADMIN PANEL */}
         <Route path="/admin-sidebar/admin-dashboard" element={<AdminDashboard />} />
         <Route path="/admin-sidebar/client-data" element={<ClientData />} />
         <Route path="/admin-sidebar/client-details" element={<ClientDetails />} />
         <Route path="/admin-sidebar/statistics" element={<Statistics />} />
 
+        {/* TESTING */}
+        <Route path='/test' element={<Test />} />
+        <Route path='/practice' element={<Practice />} />
+        <Route path='/sidebar2' element={<SideBar2 />} />
 
       </Routes>
     </>
